LayoutTests/platform/mac/accessibility/aria-listbox-selectedchildren-change-expected.txt

Option
Option
Option
This tests that when aria-selected is changed in a listbox, the correct notification is sent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
